init AppSearch
