+++ package.json
@@ -32,7 +32,9 @@
     "@nuxt-themes/typography": "^1.0.0-beta.1",
     "@nuxt/content": "^2.6.0",
     "@nuxthq/studio": "^0.13.2",
-    "@vueuse/nuxt": "^10.1.2"
+    "@vueuse/integrations": "^10.1.2",
+    "@vueuse/nuxt": "^10.1.2",
+    "fuse.js": "^6.6.2"
   },
   "devDependencies": {
     "@algolia/client-search": "^4.17.2",
